Home page random posts cards as well as the random post button made functional and NOW css

+++ public/scripts/script.js
@@ -1,3 +1,5 @@
 $(".dlt-btn").on("click", function (item) {
-  axios.delete(`/delete-post?${this.id}`);
+  if (confirm("Do you want to delete this post?")) {
+    axios.delete(`/delete-post?${this.id}`);
+  }
 });
